Refactor DKPN to avoid cross-imports
Moved shared functionality to a dedicated submodule in seisbench.util, so that no imports between seisbench.models and seisbench.generate are required.
